chore: remove api para usar a api externa escrita em go :)
