+++ .env.example
@@ -1,5 +1,4 @@
-DATABASE_URL=
-NEXT_PUBLIC_API_BASE_URL=
+RETROVISIONARIOS_API_BASE_URL=
 SERVER_API_TOKEN=
 NEXT_PUBLIC_IMAGES_HOST=
 GEMINI_API_KEY=
