+++ package.json
@@ -14,7 +14,6 @@
   "dependencies": {
     "@google/genai": "^1.30.0",
     "@google/generative-ai": "^0.24.1",
-    "@prisma/client": "^6.10.1",
     "@prisma/extension-accelerate": "^2.0.1",
     "@vercel/analytics": "^1.5.0",
     "next": "^14.2.33",
@@ -38,7 +37,6 @@
     "jest-environment-jsdom": "^30.2.0",
     "postcss": "^8",
     "prettier": "^3.6.2",
-    "prisma": "^6.10.1",
     "tailwindcss": "^3.4.1",
     "ts-jest": "^29.4.5",
     "ts-node": "^10.9.2",
